\name{aracne.networks-package}
\Rdversion{1.4}
\alias{aracne.networks-package}
\alias{aracne.networks}
\docType{package}
\title{
ARACNe-inferred gene networks from TCGA tumor datasets
}
\description{
This package contains ARACNe-inferred networks from TCGA tumor
datasets and functions to import new ones and export them into text form.}
\details{
\tabular{ll}{
Package: \tab aracne.networks\cr
Type: \tab Package\cr
Version: \tab 0.99.2\cr
Date: \tab 2016-06-07\cr
License: \tab LGPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Federico M. Giorgi

Maintainer: Federico M. Giorgi <federico.giorgi@gmail.com>
}
\references{
Giorgi,F.M. et al. (2016) ARACNe-AP: Gene Network Reverse Engineering through Adaptive Partitioning inference of Mutual Information. Bioinformatics 2016
Alvarez, M.J. et al. (2016) Functional characterization of somatic mutations in cancer using network-based inference of protein activity. Nature Genetics 2016
}
